feat: decouple openai from `llamaindex` module
